Move all LocalQueue methods to localqueue.go.
