+++ pkg/cache/scheduler/localqueue.go
@@ -21,8 +21,10 @@ import (
 
 	corev1 "k8s.io/api/core/v1"
 
+	"sigs.k8s.io/kueue/pkg/metrics"
 	"sigs.k8s.io/kueue/pkg/resources"
 	"sigs.k8s.io/kueue/pkg/util/queue"
+	"sigs.k8s.io/kueue/pkg/util/roletracker"
 )
 
 type LocalQueue struct {
@@ -34,14 +36,29 @@ type LocalQueue struct {
 	admittedUsage      resources.FlavorResourceQuantities
 }
 
-func (lq *LocalQueue) GetAdmittedUsage() corev1.ResourceList {
-	lq.RLock()
-	defer lq.RUnlock()
-	return lq.admittedUsage.FlattenFlavors().ToResourceList()
+func (q *LocalQueue) GetAdmittedUsage() corev1.ResourceList {
+	q.RLock()
+	defer q.RUnlock()
+	return q.admittedUsage.FlattenFlavors().ToResourceList()
 }
 
-func (lq *LocalQueue) updateAdmittedUsage(usage resources.FlavorResourceQuantities, op usageOp) {
-	lq.Lock()
-	defer lq.Unlock()
-	updateFlavorUsage(usage, lq.admittedUsage, op)
+func (q *LocalQueue) resetFlavorsAndResources(cqUsage resources.FlavorResourceQuantities, cqAdmittedUsage resources.FlavorResourceQuantities) {
+	// Clean up removed flavors or resources.
+	q.Lock()
+	defer q.Unlock()
+	q.totalReserved = resetUsage(q.totalReserved, cqUsage)
+	q.admittedUsage = resetUsage(q.admittedUsage, cqAdmittedUsage)
+}
+
+func (q *LocalQueue) updateAdmittedUsage(usage resources.FlavorResourceQuantities, op usageOp) {
+	q.Lock()
+	defer q.Unlock()
+	updateFlavorUsage(usage, q.admittedUsage, op)
+}
+
+func (q *LocalQueue) reportActiveWorkloads(tracker *roletracker.RoleTracker) {
+	role := roletracker.GetRole(tracker)
+	namespace, name := queue.MustParseLocalQueueReference(q.key)
+	metrics.LocalQueueAdmittedActiveWorkloads.WithLabelValues(string(name), namespace, role).Set(float64(q.admittedWorkloads))
+	metrics.LocalQueueReservingActiveWorkloads.WithLabelValues(string(name), namespace, role).Set(float64(q.reservingWorkloads))
 }
